move labels to contentVisibilityTranslations
